# services/backend_py3/swagger.yaml
openapi: "3.0.0"
info:
  description: English bot backend
  version: 1.0.0
  title: English bot backend

paths:
  /ping:
    get:
      description: Ручка с пингом
      operationId: backend_py3.src.controllers.common.ping
      responses:
        '200':
          description: OK
          content: {}

  /v1/word:
    get:
      description: Отдать все английские слова пользователю
      operationId: backend_py3.src.controllers.dictionary.get_word
      parameters:
        - in: query
          name: page
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - total_records
                  - message
                properties:
                  data:
                    type: array
                    description: Список слов
                  total_records:
                    type: integer
                    description: Количество слов всего
                  message:
                    type: string
                    description: Короткий ответ пользователю

  /v1/wordBySubject:
    get:
      description: Отдать все английские слова пользователю
      operationId: backend_py3.src.controllers.dictionary.get_word_by_subject
      parameters:
        - in: query
          name: page
          required: true
          schema:
            type: integer
        - in: query
          name: subject
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - total_records
                  - message
                properties:
                  data:
                    type: array
                    description: Список слов
                  total_records:
                    type: integer
                    description: Количество слов всего
                  message:
                    type: string
                    description: Короткий ответ пользователю

  /v1/translation:
    get:
      description: Отдать все английские слова пользователю
      operationId: backend_py3.src.controllers.dictionary.get_translation
      parameters:
        - in: query
          name: english_word
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - message
                properties:
                  data:
                    type: object
                    description: Список слов
                  message:
                    type: string
                    description: Короткий ответ пользователю

  /v1/subject:
    get:
      description: Отдать все темы пользователю
      operationId: backend_py3.src.controllers.subject.get_subject
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - message
                properties:
                  data:
                    type: array
                    description: Объект данных с темами
                  message:
                    type: string
                    description: Короткий ответ пользователю